requirements.txt
PureCloudPlatformClientV2
python-dotenv
pandas
flask
Flask-SQLAlchemy
Flask-WTF
email_validator
